requirements.txt
aiohttp
pydantic
pyyaml
pandas
numpy
